requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
python-dotenv==1.0.1
requests==2.31.0
